module and cmd files naming updated
