--- client/tracer.go
@@ -1,61 +1,0 @@
-package client
-
-import (
-	"context"
-	"log"
-
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/exporters/jaeger"
-	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
-	"go.opentelemetry.io/otel/trace"
-)
-
-func newExporter(ctx context.Context) (sdktrace.SpanExporter, error) {
-	// Your preferred exporter: console, jaeger, zipkin, OTLP, etc.
-	exp, err := jaeger.New(jaeger.WithCollectorEndpoint(jaeger.WithEndpoint("http://localhost:14268/api/traces")))
-	return exp, err
-}
-
-func newTraceProvider(exp sdktrace.SpanExporter) *sdktrace.TracerProvider {
-	// Ensure default SDK resources and the required service name are set.
-	r, err := resource.Merge(
-		resource.Default(),
-		resource.NewWithAttributes(
-			semconv.SchemaURL,
-			semconv.ServiceName("Client"),
-		),
-	)
-
-	if err != nil {
-		panic(err)
-	}
-
-	return sdktrace.NewTracerProvider(
-		sdktrace.WithBatcher(exp),
-		sdktrace.WithResource(r),
-	)
-}
-
-func CreateTracer(ctx context.Context) (*sdktrace.TracerProvider, trace.Tracer) {
-
-	exp, err := newExporter(ctx)
-	if err != nil {
-		log.Fatalf("failed to initialize exporter: %v", err)
-	}
-
-	// Create a new tracer provider with a batch span processor and the given exporter.
-	tp := newTraceProvider(exp)
-
-	// Handle shutdown properly so nothing leaks.
-	// defer func() { _ = tp.Shutdown(ctx) }()
-
-	otel.SetTracerProvider(tp)
-
-	// set global propagator to tracecontext (the default is no-op).
-	//otel.SetTextMapPropagator(propagation.TraceContext{})
-
-	// Finally, set the tracer that can be used for this package.
-	return tp, tp.Tracer("ClientService")
-}
